slight restructuring
